+++ tests/Integration/RebuildCommand/RebuildCommandTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Config\Tests\Integration\RebuildCommand;
 
-use Yiisoft\Config\Options;
+use Yiisoft\Config\Composer\Options;
 use Yiisoft\Config\Tests\Integration\IntegrationTestCase;
 
 final class RebuildCommandTest extends IntegrationTestCase